refactor(ventas): split 754-LOC page into focused module
Splits app/rdb/ventas/page.tsx into components/ventas/ following the
same convention used by app-shell/, health/, travel/, rh/, and tasks/.

Module layout:
  - types.ts          shared shapes + STATUS_OPTIONS
  - utils.ts          TZ + formatDate/formatCurrency/statusVariant + rangeForPreset
  - summary-bar.tsx   Pedidos count + Total cards
  - order-detail.tsx  drawer (items + pagos + totals + print)
  - ventas-filters.tsx  search + status + corte + date range + preset
  - ventas-table.tsx  sortable pedidos table
  - ventas-view.tsx   orchestrator (state + fetches + filter/search logic)

page.tsx becomes a thin wrapper that mounts VentasView inside
RequireAccess. Behavior preserved 1:1 — no SQL, routing, UI, or
data-fetching semantics changed. All sub-files < 325 LOC.
